Maze and Utility classes added
Added the Maze class inside the MapGenerator class.
Also added the Utility class with two simple Odd/Even recognition functions and a public struct IntVector2 used to represent the postions inside the map.
We've added a bunch of methods to the class Maze. Right now the class is able to start generating itself but it's failing in the best-first search algorithm.

Further testing is required in all the methods of the class.