Best-first search corridorDistance implemented
This new funcionality allows us to control our prefered direction when random walking the maze. The value corridorDirection can be set from 0 to 1 with 0 meaning all corridors will try to be vertical and 1 vertical. We added a range control into the Map Generator script.

Added the Utility script with some useful functions.