Improved cell type assignation process
We wrote a new function to assign the cell type for each cell in the map. Previously we had three different functions for assigning types: Chest, Trap and Walls. This was really awful performance wise because we looped through the map three times. Now we only loop once for all the assignations.

We erased the previous functions because we don't think we will be of any use. When we create the map, all the assignations for the different types need to be done once. We canno't think about a case when we need to reassign the chest or traps with the map generated.